+++ packages/server/app/controllers/events_controller.ts
@@ -37,7 +37,11 @@ export default class EventsController {
     const queryData = request.qs()
     const typeKey = queryData['typeKey']
 
-    let [query, countQuery] = buildApiQuery(Event.query(), queryData, 'events')
+    let [query, countQuery] = await buildApiQuery({
+      modelQuery: Event.query(),
+      queryData,
+      tableName: 'events',
+    })
 
     query.apply((scopes: ExtractScopes<typeof Event>) => scopes.fullEvent())