+++ packages/server/app/helpers/query_builder.ts
@@ -6,6 +6,7 @@ import Resource from '#models/resource'
 import Person from '#models/person'
 import { TransactionClientContract } from '@adonisjs/lucid/types/database'
 import Tag from '#models/tag'
+import { getIdsByDimensionQuery } from '#helpers/query_dimensions'
 
 const defaultSearchFields: Record<string, string[]> = {
   cohorts: ['name'],
@@ -26,15 +27,28 @@ function getSortSettings(queryData: Record<string, any> = {}) {
   return [sort, order]
 }
 
-export function buildApiQuery(
-  modelQuery: any,
-  queryData: Record<string, any> = { limit: 10, offset: 0 },
+export async function buildApiQuery({
+  modelQuery,
+  queryData = { limit: 10, offset: 0 },
+  tableName,
+}: {
+  modelQuery: any
+  queryData?: Record<string, any>
   tableName: string
-) {
+  defaultSearch?: string
+}) {
   let countQuery = db.from(tableName)
   const limit = queryData['limit']
   const offset = queryData['offset']
   const searchQuery = queryData['search']
+  const filters = queryData['filter']
+  const dims = queryData['dim']
+  if (filters || dims) {
+    const typeTableName = string.singular(tableName) + '_types'
+    const result = await getIdsByDimensionQuery(tableName, typeTableName, filters, dims)
+    modelQuery.andWhereIn('id', result)
+    countQuery.andWhereIn('id', result)
+  }
 
   if (limit) {
     modelQuery.limit(limit)